Geometry.Lib/Figures/Triangle.cs
using Geometry.Lib.Abstractions;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Geometry.Lib.Figures
{
    public class Triangle : ITriangle
    {
        public double SideA { get; private set; }
        public double SideB { get; private set; }
        public double SideC { get; private set; }

        public Triangle(double sideA, double sideB, double sideC)
        {
            if (sideA <= 0 || sideB <= 0 || sideC <= 0)
            {
                throw new ArgumentException("Sides must be positive values");
            }

            if (sideA + sideB <= sideC || sideA + sideC <= sideB || sideB + sideC <= sideA)
            {
                throw new ArgumentException("The given sides do not form a valid triangle");
            }

            SideA = sideA;
            SideB = sideB;
            SideC = sideC;
        }

        public double CalculateArea()
        {
            double s = (SideA + SideB + SideC) / 2;
            return Math.Sqrt(s * (s - SideA) * (s - SideB) * (s - SideC));
        }

        public bool IsRight()
        {
            double[] sides = new double[] { SideA, SideB, SideC };
            Array.Sort(sides);
            double a = sides[0];
            double b = sides[1];
            double c = sides[2];

            return Math.Abs((a * a + b * b) - (c * c)) < 1e-10;
        }
    }
}
